begin porting leaderElection code
